fixed most errors
